feat(add): Implementando casos de uso faltantes y api key de google maps con nuevas funcionalidades

+++ src/main/java/edu/dosw/rideci/domain/model/Route.java
@@ -24,7 +24,7 @@ public class Route {
 
     private double totalDistance;
 
-    private int estimatedTime;
+    private Long estimatedTime;
 
     private String polyline;
